+++ pages/folder.tsx
@@ -1,21 +1,26 @@
 import Folder from '@/containers/Folder';
-import { Link } from '@/containers/Folder/Folder.types';
+import { Link } from '@/types/Folder.types';
 import { DEFAULT_USER_ID } from '@/services/config/default';
 import fetch from '@/services/utils/fetch';
 
 export async function getServerSideProps({ query }: { query: any }) {
   const folderId = query?.folderId;
-  const { data: fetchedData } = await fetch({
-    url: `/users/${DEFAULT_USER_ID}/links`,
-    method: 'get',
-    params: { folderId },
-  });
-
-  return {
-    props: {
-      links: fetchedData.data,
-    },
-  };
+  try {
+    const { data: fetchedData } = await fetch({
+      url: `/users/${DEFAULT_USER_ID}/links`,
+      method: 'get',
+      params: { folderId },
+    });
+    return {
+      props: {
+        links: fetchedData.data,
+      },
+    };
+  } catch {
+    return {
+      notFound: true,
+    };
+  }
 }
 
 interface Props {
