--- src/containers/Folder/Folder.types.ts
@@ -1,12 +1,0 @@
-export interface Link {
-  id: number;
-  title: string;
-  url: string;
-  createdAt?: string;
-  created_at?: string;
-  description: string;
-  imageSource?: string;
-  image_source?: string;
-  folder_id?: number;
-  updated_at?: string | null;
-}
